+++ spec/support/named_external_node.rb
@@ -1,0 +1,15 @@
+class NamedExternalNode
+  include CallableTree::Node::External
+
+  def initialize(name)
+    @name = name
+  end
+
+  def identity
+    @name
+  end
+
+  def call(input = nil)
+    input
+  end
+end